Provide the SCM locator API as a Java library
