+++ java-components/build-recipes-database/pom.xml
@@ -10,11 +10,11 @@
         <version>999-SNAPSHOT</version>
     </parent>
 
-
-    <groupId>io.github.redhat-appstudio.jvmbuild</groupId>
     <artifactId>hacbs-build-recipies-database</artifactId>
-    <version>999-SNAPSHOT</version>
 
+    <properties>
+        <maven.compiler.release>11</maven.compiler.release>
+    </properties>
 
     <dependencies>
         <dependency>
@@ -22,8 +22,20 @@
             <artifactId>jackson-dataformat-yaml</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.apache.httpcomponents</groupId>
+            <artifactId>httpclient</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.apache.maven</groupId>
             <artifactId>maven-artifact</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.maven</groupId>
+            <artifactId>maven-model</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-text</artifactId>
         </dependency>
         <dependency>
             <groupId>io.quarkiverse.jgit</groupId>
