+++ java-components/build-request-processor/pom.xml
@@ -7,9 +7,7 @@
         <artifactId>jvm-build-service-parent</artifactId>
         <version>999-SNAPSHOT</version>
     </parent>
-    <groupId>io.github.redhat-appstudio.jvmbuild</groupId>
     <artifactId>hacbs-build-request-processor</artifactId>
-    <version>999-SNAPSHOT</version>
     <dependencies>
         <dependency>
             <groupId>org.apache.httpcomponents</groupId>
